Move a function to a separate file
